cookbook/pocketflow-tool-search/requirements.txt
pocketflow>=0.0.2
google-api-python-client>=2.0.0  # Google API client
google-auth>=2.0.0  # Google authentication
google-generativeai>=0.3.0  # Gemini model
openai>=1.12.0  # for search result analysis
pyyaml>=6.0.1  # for structured output
python-dotenv>=1.0.0  # for loading environment variables
beautifulsoup4>=4.12.0
requests>=2.31.0
